# internal/pb/crowd-api-v1/crowd.api.v1.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "crowd-api-v1/crowd.api.v1.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CrowdAPIV1"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/projects": {
      "post": {
        "summary": "Projects\r\n////////////////////////////////////////////////////////////////////////////////////////////////////////////////////",
        "operationId": "CrowdAPIV1_CreateProject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateProjectResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateProjectRequest"
            }
          }
        ],
        "tags": [
          "CrowdAPIV1"
        ]
      }
    },
    "/api/v1/projects/list": {
      "post": {
        "operationId": "CrowdAPIV1_ListProjects",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListProjectsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ListProjectsRequest"
            }
          }
        ],
        "tags": [
          "CrowdAPIV1"
        ]
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "operationId": "CrowdAPIV1_GetProject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetProjectResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "CrowdAPIV1"
        ]
      }
    },
    "/api/v1/tasks/resolve_by_project": {
      "post": {
        "summary": "Tasks\r\n////////////////////////////////////////////////////////////////////////////////////////////////////////////////////\r\nGet tasks by project id. This method reserves tasks for the user.",
        "operationId": "CrowdAPIV1_ResolveTasksByProject",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ResolveTasksByProjectResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ResolveTasksByProjectRequest"
            }
          }
        ],
        "tags": [
          "CrowdAPIV1"
        ]
      }
    }
  },
  "definitions": {
    "ListProjectsRequestProjectFilter": {
      "type": "object",
      "properties": {
        "creatorId": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1AnnotatorTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "projectId": {
          "type": "integer",
          "format": "int32"
        },
        "targetOverlap": {
          "type": "integer",
          "format": "int32"
        },
        "currentOverlap": {
          "type": "integer",
          "format": "int32"
        },
        "inputData": {
          "type": "string"
        },
        "outputData": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1CreateAnnotationsResponse": {
      "type": "object"
    },
    "v1CreateProjectRequest": {
      "type": "object",
      "properties": {
        "creatorId": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "instruction": {
          "type": "string"
        },
        "taskConfig": {
          "type": "string"
        },
        "targetOverlap": {
          "type": "integer",
          "format": "int32"
        },
        "tasksPerUser": {
          "type": "integer",
          "format": "int32"
        },
        "annotatorTimeLimit": {
          "type": "string"
        }
      }
    },
    "v1CreateProjectResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1GetProjectResponse": {
      "type": "object",
      "properties": {
        "project": {
          "$ref": "#/definitions/v1Project"
        }
      }
    },
    "v1ListProjectsRequest": {
      "type": "object",
      "properties": {
        "filter": {
          "$ref": "#/definitions/ListProjectsRequestProjectFilter"
        },
        "page": {
          "$ref": "#/definitions/v1Page"
        },
        "sort": {
          "$ref": "#/definitions/v1Sort"
        }
      }
    },
    "v1ListProjectsResponse": {
      "type": "object",
      "properties": {
        "projects": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Project"
          }
        },
        "page": {
          "$ref": "#/definitions/v1PageResponse"
        }
      }
    },
    "v1Page": {
      "type": "object",
      "properties": {
        "pageNum": {
          "type": "integer",
          "format": "int32"
        },
        "perPage": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1PageResponse": {
      "type": "object",
      "properties": {
        "pagesCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1Project": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "creatorId": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "instruction": {
          "type": "string"
        },
        "targetOverlap": {
          "type": "integer",
          "format": "int32"
        },
        "tasksPerUser": {
          "type": "integer",
          "format": "int32"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "annotatorTimeLimit": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "taskConfig": {
          "type": "string"
        }
      },
      "title": "MODELS"
    },
    "v1ResolveTasksByProjectRequest": {
      "type": "object",
      "properties": {
        "projectId": {
          "type": "integer",
          "format": "int32"
        },
        "userId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ResolveTasksByProjectResponse": {
      "type": "object",
      "properties": {
        "tasks": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1AnnotatorTask"
          }
        },
        "annotationDeadline": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1Sort": {
      "type": "object",
      "properties": {
        "sortField": {
          "type": "string"
        },
        "desc": {
          "type": "boolean"
        }
      }
    }
  }
}